streamlit_app/requirements.txt
altair==5.1.2
annotated-types==0.5.0
anyio==3.7.1
attrs==23.1.0
blinker==1.6.3
boto3==1.28.62
botocore==1.31.62
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.1.3
fastapi==0.103.0
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
lightgbm==4.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mysql-connector-python==8.1.0
numpy==1.25.2
packaging==23.1
pandas==2.0.3
Pillow==10.1.0
pluggy==1.3.0
protobuf==4.21.12
pyarrow==13.0.0
pydantic==2.3.0
pydantic_core==2.6.3
pydeck==0.8.1b0
Pygments==2.16.1
pytest==7.4.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.10.6
s3transfer==0.7.0
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
starlette==0.27.0
streamlit==1.28.0
tenacity==8.2.3
threadpoolctl==3.2.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.6
uvicorn==0.23.2
validators==0.22.0
watchdog==3.0.0
zipp==3.17.0
